+++ Code/LYRUIConversationListViewController.m
@@ -11,14 +11,9 @@
 #import "LYRUIDataSourceChange.h"
 #import "LYRUIConstants.h"
 
-@interface LYRUIConversationListViewController () <UISearchBarDelegate, UISearchDisplayDelegate, LYRQueryControllerDelegate>
+@interface LYRUIConversationListViewController () <LYRQueryControllerDelegate>
 
-@property (nonatomic) UISearchBar *searchBar;
-@property (nonatomic) UISearchDisplayController *searchController;
-@property (nonatomic) NSMutableArray *filteredConversations;
-@property (nonatomic) NSPredicate *searchPredicate;
 @property (nonatomic) LYRQueryController *queryController;
-@property (nonatomic) NSMutableArray *objectChages;
 @property (nonatomic) BOOL isOnScreen;
 
 @end
@@ -63,19 +58,6 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     self.title = @"Messages";
     self.accessibilityLabel = @"Messages";
     self.tableView.accessibilityLabel = @"Conversation List";
-    
-    // Search Bar Setup
-//    self.searchBar = [[UISearchBar alloc] initWithFrame:CGRectZero];
-//    self.searchBar.accessibilityLabel = @"Search Bar";
-//    self.searchBar.delegate = self;
-//    self.searchController = [[UISearchDisplayController alloc] initWithSearchBar:self.searchBar contentsController:self];
-//    self.searchController.delegate = self;
-//    self.searchController.searchResultsDelegate = self;
-//    self.searchController.searchResultsDataSource = self;
-
-    // Set Search Bar as Table View Header
-    //self.tableView.tableHeaderView = self.searchBar;
-    //[self.tableView setContentOffset:CGPointMake(0, 44)];
 
     // DataSoure
     [self setupConversationDataSource];
@@ -88,11 +70,8 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     
     // Set public configuration properties once view has loaded
     [self.tableView registerClass:self.cellClass forCellReuseIdentifier:LYRUIConversationCellReuseIdentifier];
-    [self.searchController.searchResultsTableView registerClass:self.cellClass forCellReuseIdentifier:LYRUIConversationCellReuseIdentifier];
-    
     self.tableView.rowHeight = self.rowHeight;
-    self.searchController.searchResultsTableView.rowHeight = self.rowHeight;
-    
+
     if (self.allowsEditing) {
         [self addEditButton];
     }
@@ -172,41 +151,6 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     BOOL success = [self.queryController execute:&error];
     if (!success) NSLog(@"LayerKit failed to execute query with error: %@", error);
     [self.tableView reloadData];
-}
-
-- (void)reloadConversations
-{
-    if (self.searchController.active) {
-        [self.searchController.searchResultsTableView reloadData];
-    } else {
-        [self.tableView reloadData];
-    }
-}
-
-- (BOOL)isSearching
-{
-    return self.searchController.active;
-}
-
-#pragma mark - UISearchDisplayDelegate Methods
-
-- (void)searchDisplayControllerWillBeginSearch:(UISearchDisplayController *)controller
-{
-    // We react to search begining
-}
-
-- (void)searchDisplayControllerDidEndSearch:(UISearchDisplayController *)controller
-{
-    // We respond to ending the search
-}
-
-- (void)searchBar:(UISearchBar *)searchBar textDidChange:(NSString *)searchText
-{
-    if ([self.dataSource respondsToSelector:@selector(conversationListViewController:didSearchWithString:completion:)]) {
-        [self.dataSource conversationListViewController:self didSearchWithString:searchText completion:^(NSSet *conversations) {
-            [self.tableView reloadData];
-        }];
-    }
 }
 
 #pragma mark - Table view data source methods